Update and rename PyNUT.py to PyNUT_Py3.py
PyNUT_Py3 is a Python 3 abstraction class to access NUT server(s). You can use it in Python 3 programs to access NUT's upsd data server in a simple way, without having to know the NUT protocol. To import it on Python programs you have to use the following (case sensitive) : 'import PyNUT_Py3'

This module provides a 'PyNUTClient' class that can be used to connect and get data from an upsd data server.